requirements.txt
discum==1.4.1
numpy==1.21.6
PyJWT
pyppeteer==1.0.2
pyppeteer_ghost_cursor==0.2.3
PyYAML==6.0
requests==2.28.1
user_agents==2.2.0
websocket_client==1.4.2
